MAJOR fix for some dumb reason
Fatal error on SDL_DestroyTexture() when destroying a texture that was made with a certain renderer after destroying said texture... How was this an issue before? Idk. Why does it not crash on debug mode? Idk. Why does it not even reach try catch but straight up crash? Idk. This fixes it though...